refactor: colocate files related to code block transformation
