+++ src/Layout.tsx
@@ -15,7 +15,7 @@ import { ExternalLinks } from './ExternalLinks'
 import { coseMenuModalOnMouseLeave, openMenuModal, toggleMenuModal } from './MenuModal/toggleMenuModal'
 import { MenuModal } from './MenuModal'
 import { autoScrollNav_SSR } from './autoScrollNav'
-import { initializeJsToggle_SSR } from './components/CodeSnippets/useSelectCodeLang'
+import { initializeJsToggle_SSR } from './code-block-controls/hooks/useSelectCodeLang'
 import { SearchLink } from './docsearch/SearchLink'
 import { navigate } from 'vike/client/router'
 import { css } from './utils/css'
