+++ src/code-block-controls/components/Pre.tsx
@@ -1,9 +1,7 @@
 export { Pre }
 
 import React, { ComponentPropsWithoutRef, useState } from 'react'
-/* Importing it here chokes the tests. I don't know why.
 import './Pre.css'
-//*/
 
 function Pre({ children, ...props }: ComponentPropsWithoutRef<'pre'> & { 'hide-menu'?: string }) {
   return (
