+++ src/code-block-controls/hooks/useMDXComponents.tsx
@@ -2,8 +2,8 @@ export { useMDXComponents }
 
 import React from 'react'
 import type { UseMdxComponents } from '@mdx-js/mdx'
-import { Pre } from './Pre.js'
-import { CodeSnippets } from './CodeSnippets.js'
+import { Pre } from '../components/Pre.js'
+import { CodeSnippets } from '../components/CodeSnippets.js'
 
 const useMDXComponents: UseMdxComponents = () => {
   return {
