+++ src/code-block-controls/remarkDetype.ts
@@ -4,7 +4,7 @@ import type { Root, Parent, Code } from 'mdast'
 import type { MdxJsxFlowElement } from 'mdast-util-mdx-jsx'
 import type { VFile } from '@mdx-js/mdx/internal-create-format-aware-processors'
 import { visit } from 'unist-util-visit'
-import { assertUsage } from './utils/assert.js'
+import { assertUsage } from '../utils/assert.js'
 import { parseMetaString } from './rehypeMetaToProps.js'
 import pc from '@brillout/picocolors'
 import module from 'node:module'
