+++ src/components/index.ts
@@ -8,4 +8,3 @@ export * from './HorizontalLine'
 export * from './CodeBlockTransformer'
 export * from './Comment'
 export * from './FileRemoved'
-export * from './CodeSnippets'
