+++ src/css/index.css
@@ -7,5 +7,4 @@
 @import './code.css';
 @import './table.css';
 @import './tooltip.css';
-@import '../components/Pre.css';
 @import '@docsearch/css';
