+++ src/index.ts
@@ -1,16 +1,8 @@
 /**********/
 /* PUBLIC */
 /**********/
-export {
-  CodeBlockTransformer,
-  Link,
-  RepoLink,
-  FileAdded,
-  FileRemoved,
-  ImportMeta,
-  Emoji,
-  TypescriptOnly,
-} from './components'
+export { CodeBlockTransformer, Link, RepoLink, FileAdded, FileRemoved, ImportMeta, Emoji } from './components'
+export { TypescriptOnly } from './code-block-controls/components/CodeSnippets'
 export { MenuToggle } from './Layout'
 export * from './components/Note'
 export * from './icons/index'
@@ -28,4 +20,4 @@ export { usePageContext } from './renderer/usePageContext'
 /************/
 // We provide our own `useMDXComponents()` to enable MDX component injection by setting `providerImportSource` to '@brillout/docpress'.
 // https://mdxjs.com/guides/injecting-components/
-export { useMDXComponents } from './components/useMDXComponents'
+export { useMDXComponents } from './code-block-controls/hooks/useMDXComponents'
