+++ src/vite.config.ts
@@ -7,9 +7,7 @@ import { parsePageSections } from './parsePageSections.js'
 import rehypePrettyCode from 'rehype-pretty-code'
 import remarkGfm from 'remark-gfm'
 import { transformerNotationDiff } from '@shikijs/transformers'
-import { remarkDetype } from './remarkDetype.js'
-import { rehypeMetaToProps } from './rehypeMetaToProps.js'
-import { shikiTransformerAutoLinks } from './shikiTransformerAutoLinks.js'
+import { rehypeMetaToProps, remarkDetype, shikiTransformerAutoLinks } from './code-block-controls/index.js'
 
 const root = process.cwd()
 const prettyCode = [
